Added resetCount in weak window

+++ preludecorrelator/contexthelpers/StrongWindowHelper.py
@@ -55,14 +55,14 @@ class StrongWindowHelper(ContextHelper):
         for t in range(len_timestamps-1,-1,-1):
             if now - self._timestamps[t][0] >= self._options["expire"]:
                print("I am {} : del timestamps[{}]".format(self._name, t))
-               self._timestamps[t][2].restoreAnalyzerContents(self._timestamps[t][1])
-               self.onIdmefRemoval(self._timestamps[t][1])
+               #self._timestamps[t][2].restoreAnalyzerContents(self._timestamps[t][1])
+               #self.onIdmefRemoval(self._timestamps[t][1])
                self._timestamps.pop(t)
 
         tmp_analyzer = AnalyzerContents()
         tmp_analyzer.saveAnalyzerContents(idmef)
         self._timestamps.append([now, idmef, tmp_analyzer, addAlertReference])
-        self.onIdmefAddition(idmef)
+        #self.onIdmefAddition(idmef)
     '''
     def addIdmef(self, idmef):
         now = time.time()
@@ -124,10 +124,11 @@ class StrongWindowHelper(ContextHelper):
      return False
 '''
 
-    def generateCorrelationAlert(self, send=True):
+    def generateCorrelationAlert(self, send=True, destroy=False):
         tmp_ctx = ctx_search(self._name)
-        self._ctx.destroy()
-        self.unbindContext()
+        if destroy:
+         self._ctx.destroy()
+         self.unbindContext()
         self.rst()
         if send:
             tmp_ctx.alert()
